Remove all graphics backend logic, delegating all responsibility to Gfx.
I'm not sure if this code was written before GFX was added to this project, or if GFX doesn't
provide enough controls, but GFX is written to be backend-agnostic.

See this:
https://github.com/gfx-rs/gfx/blob/master/src/lib.rs#L23-L36

and this:
https://github.com/gfx-rs/gfx/tree/master/src/backend

I'm not sure the current status of any of the backends, but it seems like a bad decision to layer
backend support ontop of GFX like the code is. This commit removes all opengl/direct3d/null cfg()
directives (fun fact, it also removes all windows only code-paths in this package) and instead
assumes GFX can handle the platform or backend specific details of OpenGL/Direct3D/NULL.

It might make sense to add a "NULL" backend to the GFX project directly.