Update gfx to master branch and use IDs for textures and meshes
Gpu resource creation is not possible in multiple threads for some gfx backends like OpenGL. These need to be allocated on the render thread, requiring to refactor the asset manager.
Now we establish a channel between the asset manager and the renderer to send resource creation commands. Depending on the backend the resource creation will be on done on the render thread or on loading the asset.