Made the assets example compile.
This moves the assets example to the new renderer as well as to the new input handler (although it does not use the new `actions` system yet). It still
crashes shortly after starting up, because of some resources/components not being registered. These should be registered outside of the example, and are
waiting for a future commit.

This removes nearly all if not all asset related functionality. Asset loading has not been implemented yet. Most notably, these changes would be required:

 - `assets::Directory` cannot be instantiated as there is no way to get an `assets::Allocator` from the `assets::Loader`.
 - Object/mesh files and texture files are waiting on PR #297

All removed code has been replaced with appropriate 'TODO's.

Additionally, this commit reformatted the example and split it up into various function, to make it easier to read. The initialisation of the lights and the camera are now fitted in their own function. The `run()` function prevents uses of `.unwrap()` and `.expect()`, and instead returns any errors to `main()` which can then print them out.