Issue #593
As a arc ball mecanism was wanted, a component indicated the target and
a system making the calculus for it was needed.

Tose two elements are now reality but need testing. One way to do so is
to provide a solid exemple of the mecanism. Also, maybe some tests could
be implemented.

Change the system to just correct distance between cam and target. The
free rotation will take care of the rest. Thanks to @jojolepro for the
heads up.

Add the example but I may not fully understand how the systems works
since it seem not to move.

Add missing systems, still no result, the scene stay still, or at least
seems to.

Get the switch back. Tomorrow I shall look into the quaternion and
determination of new position based on it.

The example now works fine.

Add better documentation.

forgot to change component name everywhere.... (name is more consistent now)

Better documentation and correct some errors in documentation

Fix typo