Issue #593
As a arc ball mecanism was wanted, a component indicated the target and
a system making the calculus for it was needed.

Tose two elements are now reality but need testing. One way to do so is
to provide a solid exemple of the mecanism. Also, maybe some tests could
be implemented.