Add config file for GitLab CI/CD
Since we now have access to the Gold level features of GitLab, this PR adds a basic config for it. Auto-mirroring is already setup in GitLab, but we need the config file in the repo, otherwise the mirroring removes it. This CI/CD file has been tested and is mostly ported from the Concourse config.