Refactor audio initialization to be more bundle-centric.
It's a bit confusing that:
 * Audio initialization requires you to handle an `Output`, even if you just use the default. Other subsystems don't seem to require this, and behave sensibly while still exposing the flexibility to pick something else.
 * `AudioListener` requires an output device while `DjSystem` doesn't seem to.
 * `AudioListener` requires you to set explicit left and right ear positions. Many times these can have a sensible default, but you're stuck researching that default when you're ready to implement audio.

This change moves device management into `AudioSystem`, with configurability via `AudioBundle`. `AudioBundle` also has a default implementation which references the default device, thus removing device management completely for the common case of opening an audio device for output.

It also moves the device out of `AudioListener` and sets default ear positions. Maybe someone might want their game to use two audio devices simultaneously, but if so then that's probably a system matter, not for components themselves to manage.