Fix position of camera in sprites_ordered example
The new renderer fixes the position of the camera in the world.
In the old renderer the y position of the camera was mirrored.
Whereas a point on (0,0,1) was 2 units away from the camera on
position (0,0,1) in the old renderer, this is now fixed in the new renderer.
A point on (0,0,1) is now two units away when the camera is positioned
at (0,0,-1)