Add pong audio subchapter
This commit adds Chapter 4.6 to the Amethyst book, which walks the
reader through how to add audio to their pong game.

This commit also removes the apology for an incomplete tutorial in the
Call for Contribution subchapter. Now that an audio subchapter has been
added, there is no need to apologize for the lack thereof.