CHANGED: renamed screen_to_world to screen_to_world_point ADDED: screen_ray
Due to the naming scheme we used, and the methods provided for screen_to_world,
it was actually incorrect for perspective cameras. Additionally, the naming convention
and data provided was insufficient really for proper 3d screen raycasting.

This adds screen_ray(), screen_to_world_point(), and a Ray struct  for their use.
This allows users to get a proper Ray from the screen, as well as get projected
points along this ray.  These functions are camera agnostic.