`FlatEncoder` added to amethyst_tiles for flat linear encoding which is optimized for space.
`TileMap` was not allocating enough space for to compensate for morton encoding alignment. This means that
all tilemap allocation must occur on a ^2 boundary. This fixes the case of assemetric tilemaps.

Fixed an iterator oppsy where the Ord was incorrect, causing the y-axis to be one tile short.

asdf

Update changelog with actual PR

asdf