Update to a newer, Rusty version of mikktspace
This fixes a link error on Windows. The version of mikktspace that was
used previously was a wrapper around C code and for some reason, Visual
Studio did not like it.

The mikktspace library has since translated the original C code to
(unsafe) Rust and changed its API in the process. This commit adapts the
code which uses mikktspace to the newest version.

Also, since mikktspace didn't have an official Rust-only release yet and
looks kind of inactive, the version specified in Cargo.toml is pinned to
the current newest commit instead of a proper version.