remove projection, put it directly into Camera
The setters and getters for parts of a projection have been removed. They were
broken, as the setters did not update the inverse matrix. If updating is
desired, a new Camera can be build, which should be about as fast.

This commit breaks auto_fov. I'll rewrite that utility in the next commit.