Move footnotes into links at appropriate places in text
Prior to this PR, this page in the book had a number of links at the bottom that appeared to be footnotes but which were rendering in such a way as to be unreadable. I fixed this by removing the footnotes entirely and adding links in the text where appropriate.