Fix buggy example code in the book
The 1.15 version of the book, chapter 10.1 'Controlling system execution -> Custom game data' contains erroneous example code that, if used, will result in runtime panics. It advocates manually creating an instance of World and passing it to the CustomGameDataBuilder. However, the ApplicationBuilder also creates an instance of World and proceeds to use that instance when creating the game. It doesn't have the resource registrations of the first instance and the game will therefore panic when it tries to look for those resources.