Update pong-tutorial-02.md
Just a minor touch-up to an import to make it a little bit easier to remember. 

Later on in the tutorial, we reference System with just `amethyst::ecs` (without the prelude). So if we're consistent about it here as well then new people who don't know any better won't get caught up on why one needs prelude and the other doesn't.