give examples their own cargo.toml
examples previously shared dev dependencies and paths with the main
amethyst crate.  this lead to bad import paths in examples, additional
joins needed on paths that made examples not easily portable. also it
made it difficult to copy examples as a starter project.

now that each example has a cargo.toml with dependencies we have a best
practice to start a new project packaged alongside each example and
users can copy an example folder in to their own new project with ease
and without editing anything.