Upgrade rodio to version 0.14.0, update API
In version 0.12.0, breaking changes were made in rodio's API. The major
one is the introduction of Streams, which replaced Devices as the main
way to play audio.

More precisely, there is an OutputStream, which can be constructed from
a default device or a Device that the user provides. This struct is
!Sync and !Send, i.e. it cannot be send across threads. There is,
however, an OutputStreamHandle which provides thread-safe access to an
audio device. Sinks are now created from those handles instead of Devices.
These changes are required to support [Android's AAudio API, which is not
completely thread-safe][aaudio].

Here are the adjustments made to this crate:
- Output now contains an OutputStreamHandle and the name of the device
  it was created with. It also provides a method to construct a Sink
  from the handle.
- AudioSink has been renamed to Sink.
- AudioSystem and DjSystem have been reworked according to the above
  changes.
- The bundles for the systems now insert at least two resources: the
  OutputStream resource, which must not be accessed by the systems, and
  the Output that is used by those systems.
- The program now panics if the system running it has no output devices
  (e.g. a Linux container). That should make identifying the problems
  with audio easier.

[aaudio]: https://developer.android.com/ndk/guides/audio/aaudio/aaudio#thread-safety